Refactor compilable_layer_indices
This refactors the value stored in the private attribute _compilable_layer_indices_tup to remove the leading '__CMPLBL__' string value. This was only used in the construction of the python tuple representation of the circuit when calling the `tup` property. Elsewhere the inclusion required awkwardly having to check the length and then slicing into the tuple to skip this string when accessing the actual indices. As such, the inclusion of this string has been added to the code for `tup`, and the remaining code has been cleaned up to match. In addition to being cleaner, this should be more performant (as we need to do fewer checks when working with the value of this object now). Additionally, this commit also switches to directly inlining access to the _compilable_layer_indices_tup instead of going through the property in a few places to reduce method call overhead.