Fix mistakes in creation of editable copy
The expected format for the labels when a circuit is editable is a nested list of lists of simple labels. When making an editable copy of a static circuit the new version of copy was not putting the labels in this format as needed. We now cache the previously editable version of the circuit layers at the point where the circuit is made static, and use this for speeding up the conversion back to an editable format.