More circuit primitive speedups and bugfixes
This commit adds the following:
A new helper function for speeding up copying by avoiding double circuit tuple construction. A new method for sandwiching a circuit with two labels (faster than is using the add operation twice). A bugfix for add which checks for the need to add to the line_labels of a circuit when the added label has a new one not presently in the circuit. Corrections and inlining of editable label to static label conversions. A slightly faster reworking of expand_subcircuits_inplace.