Initial commit porting over LFH code
This is the initial commit for the addition of functionality in pygsti for modeling systems with LFH noise present. This adds a new lfh module to extras (for now) which includes a special type of LindbladErrorgen that can have gaussianly fluctuating hamiltonian parameters. Additionally this adds a special LFH aware explicit model class, and a number of specialized forward simulators.