Add first party support for interleaved randomized benchmarking
This commit adds new functionality for performing interleaved randomized benchmarking including functionality for building the experiment designs required as well as running the protocol and storing the results into a format that is compatible with serialization and maintains ease of access to relevant subresults. This also includes a bug fix for a serialization issue that had been impacting CliffordRBDesign when using the (previously) undocumented interleaved_circuit argument.