Adds preliminary support for wildcard budgets using non-ExplicitOpModel models.
Adds alternate code pathways (as minimally as I could) for computing
wildcard budget values with a LocalNoiseModel instead of the usual
ExplicitOpModel.  This updates the wildcard analysis code in gst.py
as well as the code for computing the diamond-distance "report"
function so that the needed values can be computed.  If we refactor
the models to present a more similar interface in the future this
may be unnecessary, but using non-ExplicitOpModels is becoming more
common and computing wildcard error for such models is presently
useful.