Add circuit parameter lazifications
Add in a new method for OpModels which gets the parameter dependence for a set of circuits. This is then used in the construction of a new parameter dependence aware prefix table implementation tailored for jacobian calculations that reduces the amount of redundant computation performed. Update the generic map forward simulator calculation module to use this new (more efficient) codepath.