Revert circuit parameter dependence cython
Detailed profiling of the new parameter dependence logic in the cython code suggests that at best we break even with the older implementation using caching. As such, it is better to revert this implementation. We'll also want to detect when cython is being used to avoid building the jacobian prefix tables here, as they won't be getting used and they are an expensive upfront cost.

Keep all of the cython microoptimizations and compiler directives though, as these seem to amount to about a 10% net win once the rest of the extra logic has been stripped.