Improved performance for errorgen updates
Update the implementation of the error generator representation update code for the dense rep case. The results are functionally identical, but are measurably faster. (Einsum is ~2-3X faster than tensordot for this particular case, e.g.). We also now do the entire error generator construction in a single shot instead of block-by-block to get additional benefits from vectorization.