Further refinements to set_parameter_values
Refactor the single parameter `set_parameter_value` method to call the multi-parameter implementation under the hood.

Add additional performance tweaks to the logic for determining when to update an element of the cache. What I came up with was that when the layer rules are the ExplicitLayerRules and we are updating an effect which is known to belong to a POVM which has already been updated then we can skip the cache update for those effects.