Attempt at making PrefixTable deterministic
This commit makes changes to the implementation of the new prefix table splitting algorithm in an attempt to make it deterministic. Previously we made direct use of a number of networkx iterators which turned out to be set like, and as such they had nondeterministic order for their returned values. This is fine in the single threaded setting, but with MPI this meant we would end up with different splittings for the different ranks (I expect there to be many degenerate solutions to the splitting problem, so this isn't a crazy thing to see). This resulted in bugs when using MPI. Hopefully this should fix things...