Implement Fisher Info for Instruments
Updates the implementation of the fisher information calculation to support instruments. In addition to this there is a rework to the way that models are regularized to remove the use of depolarization and instead add in a minimum probability used for clipping small values. This was done in part to avoid the need to add in extra handling for regularizing instruments. Seems to still give sensible results for models with known fisher info spectra.

Also patches back in support for the 'cumulative' kwarg when computing fisher information by L (I could've sworn we had this implemented, but maybe it got lost at some point?).