Update unit tests
Add unit tests for instrument fisher information and modify existing unit tests. It looks like the change in the regularization likely adds some additional numerical gunk to the accumulation process which was resulting in small, but nonzero differences in the FIM matrices depending on if they were accumulated all at once, or length by length. Or at least that is my best guess for what is going on, hopefully that isn't incorrect.