Generalize CompleteElementaryErrorgenBasis
Generalize the implementation of CompleteElementaryErrorgenBasis to add support for local error generator label types. This enables constructing bases with element matrices that are given on the full space, rather than restricted to the nontrivial subspace, which is useful for some applications. Also add a new casting method for LocalStimErrorgenLabel to allow conversion from other label types/label-like objects.