Remove layerwise BCH implementation
Remove layerwise BCH implementation. There was a bug in this implementation that I couldn't track down. But more importantly, I couldn't think of any reason why this would have an advantage over propagating everything to the end before doing the pairwise recombinations (in fact, the way I just stated ought to be more performant in general).