Caching and flexible errogen label types
This commit represents an attempt at improving the performance of the construction and use of error generator coefficient dictionaries from the objects that support their creation. This is achieved two ways. First there are internal changes to the  LindbladCoefficientBlock code which caches the error generator dictionary for a block upon creation and only rebuilds this from scratch when the block data has changed.

The second change is to allow greater flexibility in the type of error generator label returned when generating the term dictionary. These can now be returned either as local or global elementary error generators, and similarly the setter methods can accept either one. This flexibility allows for performance benefits by reducing the number of casts required between error generator label types.