Fix bugs introduced by new local label handling
Fix a number of bugs and incompatibilities that were discovered related to the changes made to allow for more flexible use of either global or local elementary error generator labels. The main problem area was in the embedded op code which needed logic for dealing with embedding and unembedding error generators into larger hilbert schmidt spaces when working with local label types where basis element labels needed to be expanded or contracted.

Also fixes an issue with ComposedOp error generators related to NoErrorgenInterface inheriting objects (not sure why global error generator labels didn't previously bump into this).

Finally, add a second code path to the code for converting to stim pauli strings in the tools module to take advantage of efficiencies when using local label types.