Minor performance updates
Reduce the need to convert to local labels when we're casting from a local label. Also update code to pass in a value for the initial_label kwarg in cases where we already have a local label handy.

Improve hashing performance for LocalStimErrorgenLabel by storing a hashable representation of the pauli strings, rather than needing to recompute this every time.

Micro optimizations for the stim tableau propagation method.