Coefficient dict performance tweaks
Makes two tweaks to the implementation of the coefficient dictionary retrieval. The first moves some expensive basis handling code that is only needed when we're returning the basis for each block into a conditional block. The second reworks a loop in ComposedOp to reduce the number of dictionary look ups (and hashes).