CPTP Serialization Sign Change Fix
This fixes a bug/unexpected behavior in the serialization/deserialization of CPTP parameterized models which use the cholesky decomposition to store the values of the SCA coefficient block. Due to sign ambiguity it was the case that serializing and deserializing a model could result in a sign change for some parameters. While this wasn't necessarily wrong (both choices of sign map to the same operation), not being able to rely on these being the same could lead to weird edge cases and unexpected behavior. This addresses that by adding the parameter value attribute for the error generator to the serialized json.