Fixes bug in objective function compute_hessian routine.
Computes "local-to-current-atom" values of `firsts` (first outcome
index of a circuit with omitted outcomes) and the list of indices of
circuits with omitted outcomes.  Previously these were taken to be
global values self.firsts and self.indicesOfCircuitsWithOmittedData,
which are only correct for layouts having a *single* atom.  For
multi-atom layouts this bug would have caused index out-of-bounds
errors when trying to place omitted probability terms into the
hessian elements being assembled.