Updates list_to_slice(...) for edge case to avoid step=0 error.
Adds additional condition that the step cannot be zero when checking
whether a list of indices coincides with a slice.  Otherwise, when
given a list of all the same index, e.g. [1, 1, 1], list_to_slice
will try to compare this with a range(1,1,0) which is an error
(range does not allow step == 0). This error could occur particularly
when serializing small pygsti models that collect parameters so that
a model member's .gpindices can have repeated indices -- or even only a
single repeated index.