Fixes array-dimension bug within LindbladCoefficientBlock.superop_hessian_wrt_params()
This code hasn't been accessed much or at all recently, and it's likely
when we refactored code to create the LindbladCoefficientBlock a long
time ago this function was incorrectly updated to use the full number
of parameters in places where it should have used the square root of
this number (!).  This commit fixes this error and allows
superop_hessian_wrt_params to run at least without getting IndexErrors.