Fix additional bugs in lindbladcoefficients
Fixes a couple bugs in LindbladCoefficientBlock. The first was related to an old parameter name that was missed in a previous refactor. The second was a bug in the superop hessian calculation for the other block. Erik fixes this for cholesky parameterization, but the same fix also needed to be added for elements.

Unrelatedly, some documentation has been updated.