More processorspec cleanup (for instruments)
Continue the quest to properly document the convention expected for various non-standard operation specifiers. In this commit there are additional details added regarding instruments specifically. There are also changes made to the implementation of the instrument builder in the model construction code which fills in gaps between the supported specifier options for preps and POVM effects to ensure parity/consistency. Also does away with some limitations in the specifier length when using 2-tuples of prep and effect specifiers.