Checkpoint error generator composition implementation
Checkpoint the implementation of the pygsti code for the analytic computation of the composition of elementary error generators. This push includes H-H, H-S and H-C so far.

Additionally adds new helper/utility functions for doing numeric versions of the composition and commutator calculations (not scalable but will be useful for testing). Aside from this a few minor refactors/variable renaming.