Adds better basis resolution in opsfn_factory for LocalNoiseModels.
When a basis is needed when computing a metric on a LocalNoiseModel
the model.basis is not always the correct basis since the metric
may be on lesser-dimensional objects (e.g. 1Q metrics whe the model
and therefore model.basis is multi-qubit).  This commit tests for
this case when the Pauli-product basis is used (90+% of the time)
and constructs the appropriate reduced-dimension basis.