Temporarily disable LFH extras package
Temporarily comment out the code that is only associated with the LFH error generator and simulation extras package. It was probably a mistake to build the error generator propagation functionality on top of this branch in hindsight (but I though I needed some of the backend changes associated with this branch at the start). This will be resuscitated on a different branch when I have time to write proper tests and ensure the functionality is ready for integration.