Patch random error generator rates
Patch an issue in the random error generator construction code which had previously meant that we were getting non-CP maps when we didn't have both C and A present. It looks like with this new construction scheme we're getting CP error generators even when only using one or the other (and also when having mismatched weights).