Modify covariance specifications to include gate labels
Update the implementation of the OUCovarianceFunction to account for the gate label that gave rise to a particular error generator. As side effect of this change is that the non-Markovian simulation stuff is now temporarily restricted to ExplicitOpModel until I can figure out a way to achieve something akin to this for ImplicitOpModel.