Change LGST ckt storage with FPR
Change the way the LGST circuits are stored when using FPR. Previously these were stripped from the plaquettes and any needed circuits were added back in in _additional_circuits. Now  we just don't remove those for L=1 if we shouldn't. This addresses a bug that was present in the color box plots where the result of the old behavior was that these circuits weren't rendered with the plaquettes.