Fix strange rescaling of new per-sequence detail
A series of changes related to the rendering of the new per-sequence detail plots when using the new and improved implementation. This was tracked back to behavior related to trying to have use the aspect ratio of the figure used to rescale its size based on the dimensions of the container. This resulted in unexpected scaling behavior where really wide plots got scaled to be too small. The workaround added, since I didn't see an obvious way to address the logic, was to add the plumbing for skipping this aspect ratio locking behavior for these particular figures.