Reenable support for predicted RB number
Reenable support for the predicted RB number table in the gauge invariant error metric section of the reports. This had previously been disabled, but I believe the version of the relevant function for computing this in the pyGSTi rbtheory tools module does work, so this reenables that computation using the newest code path.