Patch bug in color interpolation
Patches a bug in the interpolation of color for SequentialColorMap which is used for TVD. This impacted the color of the new TVD histograms. Bug was related to the way that `normalize` was implemented as a pass through for this and other classes. There is a bunch of code elsewhere that relied on that particular behavior though, so I forked off a separate normalize_interpolate method to enable handling this case-by-case as needed for each ColorMap subclass (other ColorMap can be implemented once they are actually needed).