Hack for #484
This is a hacky fix for the issue reported in #484. It was reported there that when serializing and deserializing certain `PlaquetteGridCircuitStructure` objects one wouldn't get exactly identical. Drilling down it was determined that the underlying issue wasn't serialization, but in fact one could reproduce this by taking the `PlaquetteGridCircuitStructure` produced by the function `pygsti.circuits.create_lsgst_circuit_lists` (which is used in the GST experiment design construction workflow internally) and then either copy, or simply instantiate a new `PlaquetteGridCircuitStructure` pulling all of the inputs from the corresponding attributes of the originally produced one and one would get a new instance apparently identical in all ways except for the ordering of the underlying `_circuits` attribute. For the life of me I can't figure out why that is, and there is nothing obvious in the construction logic used in `pygsti.circuits.create_lsgst_circuit_lists` that I can see which would cause this. So, since based on my testing the contents are identical up to the ordering, I've introduced a hacky patch which makes a copy of the `PlaquetteGridCircuitStructure`s in `pygsti.circuits.create_lsgst_circuit_lists` before returning them.