Adds CliffordCompilationRules.add_given_clifford_compilation_template, adds iSWAP->CNOT compilation.
This new method adds a given compilation rule after checking its validity,
as an alternative to the brute force search to find such rules.  The method
is used in CliffordCompilationRules.create_standard to add a compilation
for the CNOT gate using the iSWAP gate, when required gates are available.

This commit also improves the progress printing within add_clifford_compilation_template
when verbosity > 0, so that the percentage of candidate compilations
constructed and searched is displayed as the brute-force algorithm runs.