Serialization logic for instrument specs
Add logic for serializing and deserializing newly added instrument specifications. Modify the deserialization logic in processor specifications to handle instrument specifications. This is a work in progress and will break compatibility with serialization of processor specifications that are instantiated with older style dictionary and string specifications. At least until I implement casting logic for InstrumentSpec. Also fix small bug in the deserialization of TPInstrument.