Qubit label inference for stim Tableau conversion
Generalize the method for converting a circuit to a stim Tableau to allow automatic inference of the mapping from pyGSTi labels to stim qubit labels for most standard cases. Also adds ability to specify a custom conversion behavior for cases where this doesn't work. Plus some new unit tests for the conversion methods.