Fix bug when converting unitary model to errorgen
Fix a bug in the conversion from unitary model members to error generator ones present in the new SPAM conversion logic. All that was needed here was to change the space for the to_dense calls so they were consistent. Additionally adds a docstring to convert_members_inplace and set_default_gauge_group for the explicit model class.
For set_default_gauge_group I have also changed the default behavior for the set_default_gauge_group function. Previously this would default to the TrivialGaugeGroup for all or most of the valid lindblad types. But in this case the unitary group is a better (though imperfect for edge cases) choice.