Patch wildcard for kites
Patch a bug in the computation of wildcard for kite models, or more properly just models where the pauli-product basis was not being used, by plumbing in the model's basis everywhere it needs to go.

Also fix a broken cast for kite models in compute_povm_map by switching to a different basis casting method.