Force a minimum width for table cells
This patches an issue introduced with the latest updates to the report generation code in which tables of figures would have their contents overflow and overlap because the parent table class had max width constraints which interfered with tables growing to accommodate their contents. Remove maximum width on tables, and set a minimum width on cells large enough to contain their contents.