Create new branch for development staging
This branch will serve as a staging area for bugfixes and patches intended for release as part of a bugfix release. This will enable separating these changes out from the broader class of allowed and API breaking changes that happen as part of feature development on develop.

The new workflow adds testing against pushes to this branch, and automatically merges changes on this branch into develop so that develop is always a superset of bugfix.