Enhancements and bugfixes in Report.write_notebook (#658)
This resolves issue #655.

There's a change to the public API for Report.write_notebook and changes
to the code in the notebook it generates. Report.write_notebook only
saves data as a pickle if it's called with kwarg ``use_pickle=True``.
There's a new Report.results_dict_from_dir function that can accept
_either_ the a filename with a ``.pkl`` extension or a directory name.
If the pkl extension is present then we try to load from a pickle (to
ensure backward-compatibility). Otherwise, we try to load from a
directory generated by the new Report.write_results_dict function.

I introduced a little testing infrastructure so our write_notebook test
actually runs the notebooks that it generates. I also added a test that
that calls Report.write_notebook with kwarg ``use_pickle=True``.