Remove third-order Zassenhaus
In my testing the expression for the third-order correction did not appear to result in the correct answer (with the approximation error from the third-order zassenhaus being the same or sometimes worse than second order). I was basing this off of the formulas for the generalized zassenhaus in https://arxiv.org/pdf/1903.03140. I thought I figured out what the error might have been (a missing less-than on a sum index in the formula for W3 on page 5, but that didn't seem to do the trick either). If we want to have the higher-order corrections we'll probably need to work out the correct formulas ourselves.