Deal with underflow issue
Patch an issue waiting to happen with underflowing values when dealing with very small amplitude values on states. Also fixes a related issue with the computation of phi values. This previously used the amplitudes directly, which could fail to get properly accounted for in the sign check routine I had. This was patched by adding in a specific mode for amplitude_of_state which just computes the phase of the amplitude, which is all that is needed for phi computations.