Update of Scalable Benchmarking Tool Name (B-Sqale -> scarab) (#685) (Manual Develop Merge) (#686)
This PR updates the name for the scalable algorithm benchmarking tool to
`scarab` to match the released paper (https://arxiv.org/abs/2511.02134).
The function `noise_mirror_benchmark()` has been renamed to
`lowlevel_mirror_benchmark()` to be consistent with the terminology of
the paper. This PR also moves the `scarab` utility to the
`pygsti.protocols` submodule to resolve a circular import issue.

Co-authored-by: ndsieki <ndsieki@sandia.gov>