Fix ReadTheDocs Build Failure (#691)
As was reported in issue #680, the documentation on readthedocs didn't
get built properly with the release of 0.9.14.2 a couple weeks ago.
Based on the build logs this was because the AutoAPI extension for
sphinx was failing to read the source files. Poking around I noticed
that a couple years ago I pinned the version of sphinx we were using,
apparently to address a versioning compatibility that was present at
that time. Given that I tried the easiest thing first and simply
unpinned this version in the hopes it might just be a versioning
incompatibility between that oldish version of sphinx and the newer
versions of the extensions being pulled down. Look like this did the
trick, so this PR unpins the sphinx version (whatever problem was
happening two-years ago seems to have been fixed in the intervening
time).

Also addresses a deprecation warning for a setuptools submodule in the
documentation build.

Co-authored-by: Corey Ostrove <ciostro@sandia.gov>